add authentication for two different clients & re-structure file using props
