+++ app/auth/staff/page.js
@@ -1,0 +1,12 @@
+'use client'
+import AuthForm from "../../component/authForm"
+
+export default function staffAuth(){
+    return(
+        <div>
+            <AuthForm 
+                role="Staff"
+            />
+        </div>
+    )
+}
