+++ app/auth/student/page.js
@@ -1,0 +1,12 @@
+'use client'
+import AuthForm from "../../component/authForm"
+
+export default function studentAuth(){
+    return(
+        <div>
+            <AuthForm
+                role="Student"
+            />
+        </div>
+    )
+}
